New update 3.3 (fixed and news) Create file APIConstant.java Replaced Float to Double in Entity classes Renamed api urls Fixed bugs Refactoring and optimized code Set CorsOrigin for all controllers in SecurityConfig.java

+++ src/main/java/com/max/movierating/entity/Basket.java
@@ -18,7 +18,7 @@ import java.util.Set;
 public class Basket extends BaseEntity {
 
     @ColumnDefault(value = "0")
-    private Float summa = 0.0f;
+    private Double summa = 0.0;
 
     @ManyToMany(cascade = {CascadeType.ALL})
     @JoinTable(
